refactor: extract SettingsStore+Display extensions (#636 slice A)
Move 4 read-only storage-description accessors + shared private helper
from SettingsStore (main class body) to a new SettingsStore+Display.swift
extension file. Byte-preserving move; zero production-API drift.

**Moved (68 source lines):**
- `apiKeyStorageDescription` (18 lines)
- `selectedAIProviderCredentialStorageDescription` (24 lines)
- `githubTokenStorageDescription` (6 lines)
- `jiraTokenStorageDescription` (6 lines)
- `storageDescription(for:empty:)` private helper (14 lines)

**Line-count contract met:**
- `SettingsStore.swift`: 1963 → 1895 (−68 lines)
- `SettingsStore+Display.swift`: 78 lines (< 120 target)

**Visibility unchanged:** `storageDescription(for:empty:)` stays `private`
inside the same extension file — no visibility widening at all.

**Tests added:** `SettingsStoreDisplayTests.swift` with 8 tests covering
2 persistence states (`.empty`, `.pendingSave`) × 4 accessors, using only
existing `AppStateHarness` seams. Post-refactor test count: 675 (+8).

Refined per codex spec-review (3 iterations, 12 blockers caught + fixed).

Closes #697
