+++ Sources/BugNarrator/Services/SettingsStore.swift
@@ -440,49 +440,7 @@ final class SettingsStore: ObservableObject {
         )
     }
 
-    var apiKeyStorageDescription: String {
-        if aiProvider.requiresAPIKey {
-            return storageDescription(
-                for: apiKeyPersistenceState,
-                empty: "BugNarrator never ships with an API key. Paste your own \(aiProvider.displayName) credential to enable transcription."
-            )
-        }
 
-        switch apiKeyPersistenceState {
-        case .empty:
-            return "Optional for local-compatible providers. Leave it blank if your endpoint does not require authentication."
-        default:
-            return storageDescription(
-                for: apiKeyPersistenceState,
-                empty: "Optional for local-compatible providers."
-            )
-        }
-    }
-
-    var selectedAIProviderCredentialStorageDescription: String {
-        let persistenceState = selectedAIProviderCredentialPersistenceState
-
-        if aiProvider == .parakeetLocal {
-            return "Local Parakeet does not use an API key. Check the local server connection before transcribing."
-        }
-
-        if aiProvider.requiresAPIKey {
-            return storageDescription(
-                for: persistenceState,
-                empty: "BugNarrator never ships with an API key. Paste your own \(aiProvider.displayName) credential to enable transcription."
-            )
-        }
-
-        switch persistenceState {
-        case .empty:
-            return "Optional for local-compatible providers. Leave it blank if your endpoint does not require authentication."
-        default:
-            return storageDescription(
-                for: persistenceState,
-                empty: "Optional for local-compatible providers."
-            )
-        }
-    }
 
     var openAIBaseURLValue: URL {
         Self.normalizedOpenAIBaseURL(from: openAIBaseURL, provider: aiProvider)
@@ -783,12 +741,6 @@ final class SettingsStore: ObservableObject {
         )
     }
 
-    var githubTokenStorageDescription: String {
-        storageDescription(
-            for: githubTokenPersistenceState,
-            empty: "Add a GitHub personal access token if you want to try the experimental GitHub Issues export."
-        )
-    }
 
     var normalizedGitHubRepositoryOwner: String {
         githubRepositoryOwner.trimmingCharacters(in: .whitespacesAndNewlines)
@@ -845,12 +797,6 @@ final class SettingsStore: ObservableObject {
         )
     }
 
-    var jiraTokenStorageDescription: String {
-        storageDescription(
-            for: jiraTokenPersistenceState,
-            empty: "Add Jira Cloud credentials if you want to try the experimental Jira export."
-        )
-    }
 
     var normalizedJiraBaseURL: String {
         jiraBaseURL
@@ -1654,20 +1600,6 @@ final class SettingsStore: ObservableObject {
         )
     }
 
-    private func storageDescription(for state: APIKeyPersistenceState, empty: String) -> String {
-        switch state {
-        case .empty:
-            return empty
-        case .keychain:
-            return "Stored securely in your macOS Keychain."
-        case .keychainLocked:
-            return "Stored in your macOS Keychain. BugNarrator will only prompt to unlock it when you validate the key or run an action that needs it."
-        case .sessionOnly:
-            return "Keychain storage was unavailable, so this value is only kept in memory until you quit BugNarrator."
-        case .pendingSave:
-            return "Not saved yet. BugNarrator will only prompt to use Keychain when you validate the key or run an action that needs it."
-        }
-    }
 
     private func mask(
         secret: String,
